removed un-necessary files
