--- utils/hbi-to-minikube/README.md
@@ -1,126 +1,0 @@
-# How to deploy Host Inventory to Minikube
-1.  Start `minikube` cluster:
-    ```
-    minikube start --cpus 4 --disk-size 36GB --memory 8000MB --driver=hyperkit --addons=registry --insecure-registry "10.0.0.0/24"
-    ```
-    To persist these changes for every minikube invocation, run the following:
-    ```
-    minikube config set cpus 4
-    minikube config set memory 8000
-    minikube config set disk-size 36GB
-    minikube config set vm-driver kvm2 (hyperkit for mac)
-    ```
-2.  Run `minikube ip` and note down the IP address returned.
-3.  Configure Docker Daemon to use `insecure-registry` as shown below, where `192.168.64.27` was the output of `minikube ip`:
-    ```
-    {"debug":true,"experimental":false,"insecure-registries":["127.0.0.1:5000","192.168.64.27:5000"]}
-    ```
-4.  Install the operators provided by `clowder`:
-
-    A. On a Linux system, simply run:
-    ```
-    $ curl -sS https://raw.githubusercontent.com/RedHatInsights/clowder/master/build/kube_setup.sh | bash -
-    ```
-    B. On an Apple Mac system, First download `kube_setup.sh` and insert `''` into the sed command setting namespace:
-    ```
-    $ sed -i '' "s/namespace: .*/namespace: ${STRIMZI_OPERATOR_NS}/" *RoleBinding*.yaml
-    $ bash ./kube_setup.sh
-    ```
-5.  Install clowder app:
-    ```
-    $ minikube kubectl -- apply -f https://github.com/RedHatInsights/clowder/releases/download/0.12.0/clowder-manifest-0.12.0.yaml --validate=false
-
-    # clowder pod creted in "clowder-system" namespace
-    clowder-system   clowder-controller-manager-f8f8ffb55-6xzwh   1/1   Running   0     4s
-    ```
-6.  Create a namespace for host-inventory, e.g. "hbi":
-    ```
-    $ minikube kubectl -- create namespace hbi
-    ```
-7.  Set the new namespace as the default namespace and verify.
-    ```
-    $ kubectl config set-context $(kubectl config current-context) --namespace=hbi
-    $ kubectl config view --minify=true | grep namespace
-    ```
-8.  Setup `clowder` environment:
-
-    A. Download the clowder env file:
-    ```
-    $ wget https://raw.githubusercontent.com/RedHatInsights/clowder/master/docs/examples/clowdenv.yml
-    ```
-
-    B. In the downloaderd, `clowdenv.yml`, update the documented namespace name, `jumpstart` with your namespace e.g. `hbi`:
-    ```
-    metadata:
-        name: env-hbi
-    spec:
-        targetNamespace: hbi
-    ```
-    C.  Create clowd environment:
-    ```
-    $ kubectl create -f clowdenv.yml
-    ```
-
-    The `clowdenv.yml` creates three pods:
-    ```
-    hbi   env-hbi-minio-8468b69c75-4wwxm     1/1   Running   0     26s
-    hbi   env-hbi-zookeeper-668dd5b7b-zczgt  1/1   Running   0     43s
-    hbi   env-hbi-kafka-5586985c9c-mhmld     1/1   Running   0     50s
-    ```
-9.  If not already in, go to the `host-inventory` local repo.
-10. In the HBI local repo, run:
-    ```
-    $ pipenv install --dev
-    $ pipenv shell
-    ```
-11.  Insall `crc-bonfire`:
-        ```
-        $ pip install crc-bonfire
-        ```
-12. Build `host inventory` image, push it to the minikbe `registry`, and deploy `host inventory` to `minikube` using `bonfire`.
-
-    NOTE:
-    For this write-up using MacBook Pro, "quay.io/thearifismail/ubi8-python38-psql" image was used which was built on a Red Hat registered Fedora system to execute `dnf install postgresql`.  Installing packages on to a `UBI8` requires Red Hat subscription which was not available on MacBookPro.  The Dockerfile simply copies the HBI code, upgrades pip.  When using a Red Hat system, the `dnf install` step should be in `Dockerfile` as the first step.
-13. Create a shell script, say "local_build.sh", to deploy `host-inventory` to `minikube`
-    ```
-    #!/bin/bash
-
-    set -x
-
-    IMAGE_TAG=$(git rev-parse --short=7 HEAD)
-    IMAGE="127.0.0.1:5000/host-inventory"
-
-    docker build -t $IMAGE:$IMAGE_TAG -f Dockerfile .
-    docker tag $IMAGE:$IMAGE_TAG `minikube ip`:5000/host-inventory:$IMAGE_TAG
-
-    docker push `minikube ip`:5000/host-inventory:$IMAGE_TAG
-
-    # bonfire deploy --namespace hbi -p host-inventory/host-inventory/IMAGE_TAG=$IMAGE_TAG host-inventory -i $IMAGE=$IMAGE_TAG
-
-    echo $IMAGE_TAG
-    ```
-14.  Run `local_build.sh` and wait for all pods to have started.
-15.  Make `kafka`, `host-inventory-service` containers accessible by executing the following commands in separate terminals:
-        ```
-        $ oc port-forward svc/host-inventory-service 8000
-        $ oc port-forward svc/env-hbi-kafka 29092
-        $ oc port-forward svc/host-inventory-db 5432
-        ```
-        To connect to the database, get the DB `name`, `password`, `port`, and `username` using:
-        ```
-        $ kubectl get secret --namespace=hbi host-inventory -o json > temp/hbi-app
-        $ jq -r '.data["cdappconfig.json"]' < temp/hbi-app | base64 -d > temp/host-app.json
-        ```
-16.  Set kafka topic as used by Clowder to the following, where `env-hbi` is the ClowdEnvironment name and `hbi` the namespace:
-        ```
-        HOST_INGRESS_TOPIC = os.environ.get("KAFKA_HOST_INGRESS_TOPIC", "platform.inventory.host-ingress-env-hbi-hbi")
-        ```
-17.  Create host by running:
-        ```
-        $ python utils/kafka_producer.py
-        ```
-17.  If DNS look up fails to find Kafka server, update `/etc/hosts` to link `127.0.0.1` to the Kafka server as:
-        ```
-        127.0.0.1   localhost localhost.localdomain env-hbi-kafka.hbi.svc
-        ```
-18.  Use `postman` to verify that hosts have been created.
