+++ utils/kafka_producer.py
@@ -5,7 +5,7 @@ import payloads
 from kafka import KafkaProducer
 from ttictoc import TicToc
 
-# HOST_INGRESS_TOPIC = os.environ.get("KAFKA_HOST_INGRESS_TOPIC", "platform.inventory.host-ingress-env-hbi-hbi")
+
 HOST_INGRESS_TOPIC = os.environ.get("KAFKA_HOST_INGRESS_TOPIC", "platform.inventory.host-ingress")
 BOOTSTRAP_SERVERS = os.environ.get("KAFKA_BOOTSTRAP_SERVERS", "localhost:29092")
 NUM_HOSTS = int(os.environ.get("NUM_HOSTS", 1))
